feat: Adding the bookings page and separated the dashboard into many components
